+++ src/interfaces/IRewardStreams.sol
@@ -8,7 +8,7 @@ import "./IBalanceTracker.sol";
 /// @author Euler Labs (https://www.eulerlabs.com/)
 /// @notice Interface for Reward Streams distributor contract
 interface IRewardStreams {
-    function registerReward(address rewarded, address reward, uint40 startEpoch, uint128[] calldata rewardAmounts) external;
+    function registerReward(address rewarded, address reward, uint48 startEpoch, uint128[] calldata rewardAmounts) external;
     function updateReward(address rewarded, address reward) external;
     function claimReward(address rewarded, address reward, address recipient, bool forfeitRecentReward) external;
     function claimSpilloverReward(address rewarded, address reward, address recipient) external;
@@ -18,14 +18,14 @@ interface IRewardStreams {
     function enabledRewards(address account, address rewarded) external view returns (address[] memory);
     function balanceOf(address account, address rewarded) external view returns (uint256);
     function rewardAmount(address rewarded, address reward) external view returns (uint256);
-    function rewardAmount(address rewarded, address reward, uint40 epoch) external view returns (uint256);
+    function rewardAmount(address rewarded, address reward, uint48 epoch) external view returns (uint256);
     function totalRewardedEligible(address rewarded, address reward) external view returns (uint256);
     function totalRewardRegistered(address rewarded, address reward) external view returns (uint256);
     function totalRewardClaimed(address rewarded, address reward) external view returns (uint256);
-    function currentEpoch() external view returns (uint40);
-    function getEpoch(uint40 timestamp) external view returns (uint40);
-    function getEpochStartTimestamp(uint40 epoch) external view returns (uint40);
-    function getEpochEndTimestamp(uint40 epoch) external view returns (uint40);
+    function currentEpoch() external view returns (uint48);
+    function getEpoch(uint48 timestamp) external view returns (uint48);
+    function getEpochStartTimestamp(uint48 epoch) external view returns (uint48);
+    function getEpochEndTimestamp(uint48 epoch) external view returns (uint48);
 }
 
 /// @title IStakingFreeRewardStreams
