+++ test/harness/StakingFreeRewardStreamsHarness.sol
@@ -8,9 +8,9 @@ contract StakingFreeRewardStreamsHarness is StakingFreeRewardStreams {
     using SafeERC20 for IERC20;
     using Set for SetStorage;
 
-    constructor(IEVC evc, uint40 epochDuration) StakingFreeRewardStreams(evc, epochDuration) {}
+    constructor(IEVC evc, uint48 epochDuration) StakingFreeRewardStreams(evc, epochDuration) {}
 
-    function setDistributionAmount(address rewarded, address reward, uint40 epoch, uint128 amount) external {
+    function setDistributionAmount(address rewarded, address reward, uint48 epoch, uint128 amount) external {
         distributionAmounts[rewarded][reward][_storageIndex(epoch)][_epochIndex(epoch)] = amount;
     }
