+++ test/harness/StakingRewardStreamsHarness.sol
@@ -8,9 +8,9 @@ contract StakingRewardStreamsHarness is StakingRewardStreams {
     using SafeERC20 for IERC20;
     using Set for SetStorage;
 
-    constructor(IEVC evc, uint40 epochDuration) StakingRewardStreams(evc, epochDuration) {}
+    constructor(IEVC evc, uint48 epochDuration) StakingRewardStreams(evc, epochDuration) {}
 
-    function setDistributionAmount(address rewarded, address reward, uint40 epoch, uint128 amount) external {
+    function setDistributionAmount(address rewarded, address reward, uint48 epoch, uint128 amount) external {
         distributionAmounts[rewarded][reward][_storageIndex(epoch)][_epochIndex(epoch)] = amount;
     }
